sql 쿼리문 출력 오류 문제 수정보류 해당 문장은 일단 제거함

+++ src/net/account/db/AccountDAO.java
@@ -214,7 +214,7 @@ public class AccountDAO
 			pstmt = con.prepareStatement(sql);
 			pstmt.setLong(1, account.getEXP_SQ());
 			pstmt.setLong(2, account.getUSER_NO());
-			pstmt.setInt(3, account.getAMOUNTS());
+			pstmt.setLong(3, account.getAMOUNTS());
 			pstmt.setString(4, account.getMEMO());
 			pstmt.setLong(5, account.getACCOUNTNO());
 			pstmt.setLong(6, account.getCATEGORYNO());
